Menus added, messy switches.

+++ Category.java
@@ -18,8 +18,8 @@ public class Category {
         this.funds = BigDecimal.valueOf(funds);
     }
 
-    public void addFunds(BigDecimal amountToAdd){
-        this.funds = this.funds.add(amountToAdd);
+    public void addFunds(double amountToAdd){
+        this.funds = this.funds.add(BigDecimal.valueOf(amountToAdd));
     }
 
     public void removeFunds(double amountToRemove){
@@ -30,12 +30,12 @@ public class Category {
         return this.funds;
     }
 
-    public void setFunds(BigDecimal funds){
-        this.funds = funds;
+    public void setFunds(double funds){
+        this.funds = BigDecimal.valueOf(funds);
     }
 
-    public void moveFunds(Category categoryToAddTo, BigDecimal amountToMove){
-        this.funds = this.funds.subtract(amountToMove);
+    public void moveFunds(Category categoryToAddTo, double amountToMove){
+        this.funds = this.funds .subtract(BigDecimal.valueOf(amountToMove));
         categoryToAddTo.addFunds(amountToMove);
     }
